Support copy/paste on systems that do not support primary selection
Before this change, eflex assumed that the underlying platform supports
the PRIMARY selection. In practise this means that on systems that
support it, one can use the mouse selection and middle mouse button
click to copy/paste values in and out from eflex. On systems that do not
support that however, (eg. macOS Monterey) the behavior falls back to
using ctrl/cmd+c/v (or a copy/paste menu).

This will try first primary=true and if that fails, primary=false for
the usePrimarySelection.